Allow creation of DenseVector in CRFModel, AbstractLinearSGD{Model,Trainer}
- CRFModel, AbstractLinearSGDModel and AbstractLinearSGDTrainer now use
DenseVector for the feature vector if the feature vector is the same
size as the number of features. This should optimise things which
operate on dense embedding vectors. In the future this may be converted
into a user controlled parameter which uses a DenseVector if at least
some fraction of the features are present.