Refactoring CSVLoader so it uses CSVDataSource.
This refactor is to make it easier to reproduce a model which was
trained using data loaded via CSVLoader. Previously as it produced
ListDataSource with a specific provenance then automatic reproduction
required special casing everything which used CSVLoader. Now it is
refactored to sit on top of RowProcessor and produce a regular
CSVDataSource.

There are two breaking changes:
- CSVLoader.loadDataSource now returns DataSource<T> (an instance of
  CSVDataSource<T>) instead of ListDataSource<T>.
- CSVLoader.loadDataSource is now lazy and does not cache the CSV rows.
  This causes the parsing exceptions to be thrown when the source is
iterated rather than when loadDataSource is called.