Makes LibSVM training reproducible.
LibSVM has a global RNG which is used when making probailistic
predictions. This commit uses Tribuo's standard mechanisms to track RNG
state to ensure this is reproducible, at the cost of sequentializing all
LibSVM training runs to ensure that the RNG state is consistent with the
provenance.

Also added some more Javadoc.